all tests fixed
