+++ requests/new_course.rest
@@ -1,4 +1,4 @@
-POST http://localhost:3003/api/courses
+POST http://localhost:3004/api/courses
 Content-Type: application/json
 
 {
